requirements.txt
streamlit
cv2
numpy
keras
tensorflow
matplotlib
io
pillow